+++ src/utils/pathUtils.ts
@@ -20,8 +20,6 @@ export async function getTargetPath(Uri?: vscode.Uri): Promise<{ targetPath: str
         if (parentFolderUri?.[0]) {
             targetPath = parentFolderUri[0].fsPath;
             createNewFolder = true;
-        } else {
-            vscode.window.showErrorMessage('No folder selected', { modal: true });
         }
     }
     return { targetPath, createNewFolder };